updating audiodb call from api routes
